pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tim-urban-research-agent"
version = "0.1.0"
description = "An AI research agent that creates Tim Urban-style blog posts with stick figure cartoons"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "anthropic>=0.8.0",
    "openai>=1.0.0",
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "youtube-transcript-api>=0.6.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
tim-urban-agent = "tim_urban_agent.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100
